refactor: moved api endpoints constant to lib
